Rewrite fr parser
