+++ solution/parser/ecfr-parser/main_test.go
@@ -299,167 +299,6 @@ func TestParseTitles(t *testing.T) {
 	}
 }
 
-func TestProcessPartsList(t *testing.T) {
-	testTable := []struct {
-		Name                       string
-		Title                      int
-		RawParts                   []*eregs.PartConfig
-		ExtractSubchapterPartsFunc func(context.Context, int, *ecfr.SubchapterOption) ([]string, error)
-		Output                     []*eregs.PartConfig
-		Error                      bool
-	}{
-		{
-			Name:  "test-part",
-			Title: 42,
-			RawParts: []*eregs.PartConfig{
-				&eregs.PartConfig{
-					Type:            "part",
-					Value:           "400",
-					UploadLocations: true,
-					UploadRegText:   true,
-				},
-			},
-			ExtractSubchapterPartsFunc: func(ctx context.Context, title int, subc *ecfr.SubchapterOption) ([]string, error) {
-				return []string{}, nil
-			},
-			Output: []*eregs.PartConfig{
-				&eregs.PartConfig{
-					Type:            "part",
-					Value:           "400",
-					UploadLocations: true,
-					UploadRegText:   true,
-				},
-			},
-			Error: false,
-		},
-		{
-			Name:  "test-subchapter",
-			Title: 42,
-			RawParts: []*eregs.PartConfig{
-				&eregs.PartConfig{
-					Type:            "part",
-					Value:           "399",
-					UploadLocations: true,
-					UploadRegText:   true,
-				},
-				&eregs.PartConfig{
-					Type:            "subchapter",
-					Value:           "IV-C",
-					UploadLocations: true,
-					UploadRegText:   true,
-				},
-			},
-			ExtractSubchapterPartsFunc: func(ctx context.Context, title int, subc *ecfr.SubchapterOption) ([]string, error) {
-				if subc.Chapter == "IV" && subc.Subchapter == "C" && title == 42 {
-					return []string{"400", "401", "402"}, nil
-				} else {
-					return nil, fmt.Errorf("invalid title or subchapter")
-				}
-			},
-			Output: []*eregs.PartConfig{
-				&eregs.PartConfig{
-					Type:            "part",
-					Value:           "399",
-					UploadLocations: true,
-					UploadRegText:   true,
-				},
-				&eregs.PartConfig{
-					Type:            "part",
-					Value:           "400",
-					UploadLocations: true,
-					UploadRegText:   true,
-				},
-				&eregs.PartConfig{
-					Type:            "part",
-					Value:           "401",
-					UploadLocations: true,
-					UploadRegText:   true,
-				},
-				&eregs.PartConfig{
-					Type:            "part",
-					Value:           "402",
-					UploadLocations: true,
-					UploadRegText:   true,
-				},
-			},
-			Error: false,
-		},
-		{
-			Name:  "test-extract-subchapter-parts-fail",
-			Title: 42,
-			RawParts: []*eregs.PartConfig{
-				&eregs.PartConfig{
-					Type:            "part",
-					Value:           "399",
-					UploadLocations: true,
-					UploadRegText:   true,
-				},
-				&eregs.PartConfig{
-					Type:            "subchapter",
-					Value:           "IV-C",
-					UploadLocations: true,
-					UploadRegText:   true,
-				},
-			},
-			ExtractSubchapterPartsFunc: func(ctx context.Context, title int, subc *ecfr.SubchapterOption) ([]string, error) {
-				return nil, fmt.Errorf("oops")
-			},
-			Output: nil,
-			Error:  true,
-		},
-		{
-			Name:  "test-invalid-type",
-			Title: 42,
-			RawParts: []*eregs.PartConfig{
-				&eregs.PartConfig{
-					Type:            "part",
-					Value:           "399",
-					UploadLocations: true,
-					UploadRegText:   true,
-				},
-				&eregs.PartConfig{
-					Type:            "something_else",
-					Value:           "IV-C",
-					UploadLocations: true,
-					UploadRegText:   true,
-				},
-			},
-			ExtractSubchapterPartsFunc: func(ctx context.Context, title int, subc *ecfr.SubchapterOption) ([]string, error) {
-				return []string{}, nil
-			},
-			Output: nil,
-			Error:  true,
-		},
-		{
-			Name:     "test-no-parts",
-			Title:    42,
-			RawParts: []*eregs.PartConfig{},
-			ExtractSubchapterPartsFunc: func(ctx context.Context, title int, subc *ecfr.SubchapterOption) ([]string, error) {
-				return []string{}, nil
-			},
-			Output: nil,
-			Error:  true,
-		},
-	}
-
-	for _, tc := range testTable {
-		Run(t, tc.Name, func(t *testing.T) {
-			ExtractSubchapterPartsFunc = tc.ExtractSubchapterPartsFunc
-			ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
-			defer cancel()
-			output, err := processPartsList(ctx, tc.Title, tc.RawParts)
-			diff := deep.Equal(tc.Output, output)
-			if err != nil && !tc.Error {
-				t.Errorf("expected no error, received (%+v)", err)
-			} else if err == nil && tc.Error {
-				t.Errorf("expected error, received (%+v)", output)
-			} else if err == nil && diff != nil {
-				t.Errorf("output not as expected: %+v", diff)
-			}
-		})
-	}
-}
-
 func TestParseTitle(t *testing.T) {
 	SleepFunc = func(t time.Duration) {}
 	PostParserResultFunc = func(ctx context.Context, result *eregs.ParserResult) (int, error) { return 200, nil }
