+++ solution/parser/lib/eregs/config.go
@@ -5,8 +5,10 @@ import (
 	"encoding/json"
 	"fmt"
 	"time"
+	"strings"
 
 	"github.com/cmsgov/cmcs-eregulations/lib/network"
+	"github.com/cmsgov/cmcs-eregulations/lib/ecfr"
 
 	log "github.com/sirupsen/logrus"
 )
@@ -78,4 +80,39 @@ func GetLogLevel(l string) log.Level {
 		log.Warn("[main] '", l, "' is an invalid log level, defaulting to 'warn'.")
 		return log.WarnLevel
 	}
+}
+
+var extractSubchapterPartsFunc = ecfr.ExtractSubchapterParts
+
+// ProcessPartsList converts parts and subchapters into just parts
+func ProcessPartsList(ctx context.Context, title int, rawParts []*PartConfig) ([]*PartConfig, error) {
+	var parts []*PartConfig
+	for _, part := range rawParts {
+		if part.Type == "part" {
+			parts = append(parts, part)
+		} else if part.Type == "subchapter" {
+			log.Debug("[main] Fetching parts list for title ", title, " subchapter", part.Value)
+			subchapter := strings.Split(part.Value, "-")
+			subchapterParts, err := extractSubchapterPartsFunc(ctx, title, &ecfr.SubchapterOption{subchapter[0], subchapter[1]})
+			if err != nil {
+				return nil, err
+			}
+			for _, subchapterPart := range subchapterParts {
+				parts = append(parts, &PartConfig{
+					Type:            "part",
+					Value:           subchapterPart,
+					UploadLocations: part.UploadLocations,
+					UploadRegText:   part.UploadRegText,
+				})
+			}
+		} else {
+			return nil, fmt.Errorf("%s is not a valid type for %s", part.Type, part.Value)
+		}
+	}
+
+	if len(parts) < 1 {
+		return nil, fmt.Errorf("some number of parts must be specified")
+	}
+
+	return parts, nil
 }
